FizzBuzz
The FizzBuzz program prints numbers from 1 to 100. 
If the number is the multiple of two, the program prints "Fizz". 
If the number is the multiple of seven, the program prints "Buzz".
If the number is the multiple of both two and seven, the program prints "FizzBuzz".
If the number is not the multiple of two or seven, the program prints the number.